Enhance cross-platform build and documentation for BennuGD64

+++ core/bgdrtm/src/sysprocs.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -42,9 +42,9 @@
 /* ---------------------------------------------------------------------- */
 
 /* Debe existir un header bgdrtm.h */
-extern int bgd_copy_struct( INSTANCE * my, int * params ) ;
-extern int bgd_internal_memcopy( INSTANCE * my, int * params ) ;
-extern int bgd_internal_copy_string_array( INSTANCE * my, int * params ) ;
+extern int bgd_copy_struct( INSTANCE * my, intptr_t * params ) ;
+extern int bgd_internal_memcopy( INSTANCE * my, intptr_t * params ) ;
+extern int bgd_internal_copy_string_array( INSTANCE * my, intptr_t * params ) ;
 
 #include "sysprocs.h"
 #include "sysprocs_p.h"
