+++ modules/libbgload/bgload.h
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -29,6 +29,8 @@
 #ifndef BGLOAD_H
 #define BGLOAD_H
 
+#include <stdint.h>
+
 /* --------------------------------------------------------------------------- */
 
 typedef struct
@@ -39,7 +41,7 @@ typedef struct
 
 /* --------------------------------------------------------------------------- */
 
-extern int bgload( int ( *fn )(), int * params );
+extern int bgload( int ( *fn )(), intptr_t * params );
 
 /* --------------------------------------------------------------------------- */
