+++ modules/mod_dir/mod_dir.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -78,7 +78,7 @@ enum
 
 DLVARFIXUP __bgdexport( mod_dir, globals_fixup)[] =
     {
-        /* Nombre de variable global, puntero al dato, tamaño del elemento, cantidad de elementos */
+        /* Nombre de variable global, puntero al dato, tamaÃ±o del elemento, cantidad de elementos */
         { "fileinfo.path" , NULL, -1, -1 },
         { "fileinfo.name" , NULL, -1, -1 },
         { "fileinfo.directory" , NULL, -1, -1 },
@@ -95,7 +95,7 @@ DLVARFIXUP __bgdexport( mod_dir, globals_fixup)[] =
 /* ----------------------------------------------------------------- */
 /* DIRECTORY FUNCTIONS */
 
-static int moddir_cd( INSTANCE * my, int * params )
+static int moddir_cd( INSTANCE * my, intptr_t * params )
 {
     char * d = dir_current() ;
     int r = string_new( d ) ;
@@ -104,7 +104,7 @@ static int moddir_cd( INSTANCE * my, int * params )
     return r ;
 }
 
-static int moddir_chdir( INSTANCE * my, int * params )
+static int moddir_chdir( INSTANCE * my, intptr_t * params )
 {
     const char * d = string_get( params[ 0 ] ) ;
     int ret = dir_change( d ) ;
@@ -112,7 +112,7 @@ static int moddir_chdir( INSTANCE * my, int * params )
     return ( ret ) ;
 }
 
-static int moddir_mkdir( INSTANCE * my, int * params )
+static int moddir_mkdir( INSTANCE * my, intptr_t * params )
 {
     const char * d = string_get( params[ 0 ] ) ;
     int ret = dir_create( d ) ;
@@ -120,7 +120,7 @@ static int moddir_mkdir( INSTANCE * my, int * params )
     return ( ret ) ;
 }
 
-static int moddir_rmdir( INSTANCE * my, int * params )
+static int moddir_rmdir( INSTANCE * my, intptr_t * params )
 {
     const char * d = string_get( params[ 0 ] ) ;
     int ret = dir_delete( d );
@@ -128,7 +128,7 @@ static int moddir_rmdir( INSTANCE * my, int * params )
     return ( ret ) ;
 }
 
-static int moddir_rm( INSTANCE * my, int * params )
+static int moddir_rm( INSTANCE * my, intptr_t * params )
 {
     const char * d = string_get( params[ 0 ] ) ;
     int ret = dir_deletefile( d );
@@ -200,7 +200,7 @@ static int __moddir_read(__DIR_ST * dh )
  *  until no more files exists. It then returns NIL.
  */
 
-static int moddir_glob( INSTANCE * my, int * params )
+static int moddir_glob( INSTANCE * my, intptr_t * params )
 {
     const char * path = string_get( params[ 0 ] );
     static __DIR_ST * dh = NULL;
@@ -232,7 +232,7 @@ static int moddir_glob( INSTANCE * my, int * params )
  *  return 0 if fail.
  */
 
-static int moddir_open( INSTANCE * my, int * params )
+static int moddir_open( INSTANCE * my, intptr_t * params )
 {
     int result = ( int ) dir_open( string_get( params[ 0 ] ) );
     string_discard( params[ 0 ] );
@@ -242,7 +242,7 @@ static int moddir_open( INSTANCE * my, int * params )
 /*  int DIRCLOSE (INT handle)
  */
 
-static int moddir_close( INSTANCE * my, int * params )
+static int moddir_close( INSTANCE * my, intptr_t * params )
 {
     if ( params[ 0 ] ) dir_close ( ( __DIR_ST * ) params[ 0 ] ) ;
     return 1;
@@ -255,7 +255,7 @@ static int moddir_close( INSTANCE * my, int * params )
  *  until no more files exists. It then returns NIL.
  */
 
-static int moddir_read( INSTANCE * my, int * params )
+static int moddir_read( INSTANCE * my, intptr_t * params )
 {
     return ( __moddir_read((__DIR_ST *) params[ 0 ] ) ) ;
 }
