+++ modules/mod_key/mod_key.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -77,7 +77,7 @@ static _inline int _get_key( int code )
 
 /* --------------------------------------------------------------------------- */
 
-static int modkey_key( INSTANCE * my, int * params )
+static int modkey_key( INSTANCE * my, intptr_t * params )
 {
     return ( _get_key( params[0] ) );
 }
