+++ modules/mod_string/mod_string.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -41,7 +41,7 @@
  *  Returns the size of a string
  */
 
-static int modstring_strlen( INSTANCE * my, int * params )
+static int modstring_strlen( INSTANCE * my, intptr_t * params )
 {
     const char * str = string_get( params[0] ) ;
     int r = str ? strlen( str ) : 0 ;
@@ -53,7 +53,7 @@ static int modstring_strlen( INSTANCE * my, int * params )
  *  Converts a string to upper-case
  */
 
-static int modstring_strupper( INSTANCE * my, int * params )
+static int modstring_strupper( INSTANCE * my, intptr_t * params )
 {
     int r = string_ucase( params[0] ) ;
     string_discard( params[0] ) ;
@@ -65,7 +65,7 @@ static int modstring_strupper( INSTANCE * my, int * params )
  *  Converts a string to lower-case
  */
 
-static int modstring_strlower( INSTANCE * my, int * params )
+static int modstring_strlower( INSTANCE * my, intptr_t * params )
 {
     int r = string_lcase( params[0] ) ;
     string_discard( params[0] ) ;
@@ -77,7 +77,7 @@ static int modstring_strlower( INSTANCE * my, int * params )
  *  Compares two strings, case-insensitive
  */
 
-static int modstring_strcasecmp( INSTANCE * my, int * params )
+static int modstring_strcasecmp( INSTANCE * my, intptr_t * params )
 {
     int r = string_casecmp( params[0], params[1] ) ;
     string_discard( params[0] ) ;
@@ -90,7 +90,7 @@ static int modstring_strcasecmp( INSTANCE * my, int * params )
  *  and returning a string limited to COUNT characters
  */
 
-static int modstring_substr( INSTANCE * my, int * params )
+static int modstring_substr( INSTANCE * my, intptr_t * params )
 {
     int r = string_substr( params[0], params[1], ( params[2] < 0 ) ? ( params[2] - 1 ) : params[2] ) ;
     string_discard( params[0] ) ;
@@ -102,7 +102,7 @@ static int modstring_substr( INSTANCE * my, int * params )
  *  Returns a substring, from the character given to the end of the source string
  */
 
-static int modstring_substr2( INSTANCE * my, int * params )
+static int modstring_substr2( INSTANCE * my, intptr_t * params )
 {
     int r = string_substr( params[0], params[1], -1 ) ;
     string_discard( params[0] ) ;
@@ -114,7 +114,7 @@ static int modstring_substr2( INSTANCE * my, int * params )
  *  Searchs a substring in a string, and returns its position
  */
 
-static int modstring_strfind( INSTANCE * my, int * params )
+static int modstring_strfind( INSTANCE * my, intptr_t * params )
 {
     int r = string_find( params[0], params[1], 0 ) ;
     string_discard( params[0] ) ;
@@ -126,7 +126,7 @@ static int modstring_strfind( INSTANCE * my, int * params )
  *  Searchs a substring in a string, starting from the given position, and returns its position
  */
 
-static int modstring_strfindSSI( INSTANCE * my, int * params )
+static int modstring_strfindSSI( INSTANCE * my, intptr_t * params )
 {
     int r = string_find( params[0], params[1], params[2] ) ;
     string_discard( params[0] ) ;
@@ -138,7 +138,7 @@ static int modstring_strfindSSI( INSTANCE * my, int * params )
  *  Expands the string up to the given length, adding spaces at the left
  */
 
-static int modstring_lpad( INSTANCE * my, int * params )
+static int modstring_lpad( INSTANCE * my, intptr_t * params )
 {
     int r = string_pad( params[0], params[1], 0 );
     string_discard( params[0] );
@@ -150,7 +150,7 @@ static int modstring_lpad( INSTANCE * my, int * params )
  *  Expands the string up to the given length, adding spaces at the right
  */
 
-static int modstring_rpad( INSTANCE * my, int * params )
+static int modstring_rpad( INSTANCE * my, intptr_t * params )
 {
     int r = string_pad( params[0], params[1], 1 );
     string_discard( params[0] );
@@ -162,7 +162,7 @@ static int modstring_rpad( INSTANCE * my, int * params )
  *  Converts an integer to string
  */
 
-static int modstring_itos( INSTANCE * my, int * params )
+static int modstring_itos( INSTANCE * my, intptr_t * params )
 {
     int r = string_itoa( params[0] ) ;
     string_use( r ) ;
@@ -173,7 +173,7 @@ static int modstring_itos( INSTANCE * my, int * params )
  *  Converts a floating-point number to string
  */
 
-static int modstring_ftos( INSTANCE * my, int * params )
+static int modstring_ftos( INSTANCE * my, intptr_t * params )
 {
     int r = string_ftoa( *( float * ) & params[0] ) ;
     string_use( r ) ;
@@ -184,7 +184,7 @@ static int modstring_ftos( INSTANCE * my, int * params )
  *  Converts a string to integer
  */
 
-static int modstring_stoi( INSTANCE * my, int * params )
+static int modstring_stoi( INSTANCE * my, intptr_t * params )
 {
     const char * str = string_get( params[0] ) ;
     int r = str ? atoi( str ) : 0 ;
@@ -196,7 +196,7 @@ static int modstring_stoi( INSTANCE * my, int * params )
  *  Converts a string to floating-point number
  */
 
-static int modstring_stof( INSTANCE * my, int * params )
+static int modstring_stof( INSTANCE * my, intptr_t * params )
 {
     const char * str = string_get( params[0] ) ;
     float res = ( float )( str ? atof( str ) : 0 );
@@ -208,7 +208,7 @@ static int modstring_stof( INSTANCE * my, int * params )
  *  Return the ASCII code of the first character at the string
  */
 
-static int modstring_asc( INSTANCE * my, int * params )
+static int modstring_asc( INSTANCE * my, intptr_t * params )
 {
     const unsigned char * str = ( unsigned char * ) string_get( params[0] ) ;
     int r = str ? *str : '\0' ;
@@ -220,7 +220,7 @@ static int modstring_asc( INSTANCE * my, int * params )
  *  Returns a string of length 1, with the character of the given ASCII code
  */
 
-static int modstring_chr( INSTANCE * my, int * params )
+static int modstring_chr( INSTANCE * my, intptr_t * params )
 {
     unsigned char buffer[2] = " " ; int r ;
     buffer[0] = ( unsigned char ) params[0] ;
@@ -233,7 +233,7 @@ static int modstring_chr( INSTANCE * my, int * params )
  *  Returns the given string, stripping any space characters at the beginning or the end
  */
 
-static int modstring_trim( INSTANCE * my, int * params )
+static int modstring_trim( INSTANCE * my, intptr_t * params )
 {
     int r = string_strip( params[0] ) ;
     string_discard( params[0] ) ;
@@ -265,7 +265,7 @@ static char * strrev( char * str )
  *  Returns the reverse of the source string
  */
 
-static int modstring_strrev( INSTANCE * my, int * params )
+static int modstring_strrev( INSTANCE * my, intptr_t * params )
 {
     int r = string_new( string_get( params[0] ) );
     string_discard( params[0] ) ;
@@ -278,7 +278,7 @@ static int modstring_strrev( INSTANCE * my, int * params )
  *  Converts a given integer value to string form
  */
 
-static int modstring_formatI( INSTANCE * my, int * params )
+static int modstring_formatI( INSTANCE * my, intptr_t * params )
 {
     int r = string_format( params[0], 0, '.', ',' );
     string_use( r ) ;
@@ -289,7 +289,7 @@ static int modstring_formatI( INSTANCE * my, int * params )
  *  Converts a given value to string form
  */
 
-static int modstring_formatF( INSTANCE * my, int * params )
+static int modstring_formatF( INSTANCE * my, intptr_t * params )
 {
     int r = string_format( *( float * ) & params[0], -1, '.', ',' );
     string_use( r ) ;
@@ -301,7 +301,7 @@ static int modstring_formatF( INSTANCE * my, int * params )
  *  of decimals, as given with the second parameter.
  */
 
-static int modstring_formatFI( INSTANCE * my, int * params )
+static int modstring_formatFI( INSTANCE * my, intptr_t * params )
 {
     int r = string_format( *( float * ) & params[0], params[1], '.', ',' );
     string_use( r ) ;
