+++ modules/mod_sys/mod_sys.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -33,9 +33,8 @@
 #include "files.h"
 #include "xstrings.h"
 
-#include <unistd.h>
-
 #ifndef WIN32
+#include <unistd.h>
 #include <sys/wait.h>
 #else
 #include <process.h>
@@ -47,7 +46,7 @@
 
 /* ---------------------------------------------------------------------- */
 
-static int modsys_exec( INSTANCE * my, int * params )
+static int modsys_exec( INSTANCE * my, intptr_t * params )
 {
     int mode = params[0];
     char * filename = ( char * ) string_get( params[1] );
@@ -107,7 +106,7 @@ static int modsys_exec( INSTANCE * my, int * params )
 
 /* ---------------------------------------------------------------------- */
 
-static int modsys_getenv( INSTANCE * my, int * params )
+static int modsys_getenv( INSTANCE * my, intptr_t * params )
 {
     char *e ;
     int str ;
